refactor(error): move errors into monolithic error module
