+++ src/eval.rs
@@ -1,10 +1,8 @@
 pub mod dispatcher;
-pub mod errors;
 mod parser;
 pub mod readline;
 mod symbols;
 mod tokenizer;
 
 pub use dispatcher::Dispatcher;
-pub use errors::DispatchError;
 pub use readline::LineEditor;
