+++ src/exec.rs
@@ -1,4 +1,3 @@
 mod builtin_arguments;
 pub mod builtins;
 pub mod commands;
-mod errors;
