+++ src/state.rs
@@ -1,5 +1,4 @@
 mod config;
-mod environment;
-pub mod errors;
+pub mod environment;
 pub mod path;
 pub mod shell;
